requirements.txt
alembic
beautifulsoup4
certifi
chardet
click
Flask
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
gunicorn
html5lib
idna
itsdangerous
Jinja2
lxml
Mako
MarkupSafe
mysql-connector
python-dateutil
python-dotenv
python-editor
requests
six
soupsieve
SQLAlchemy
urllib3
webencodings
Werkzeug
WTForms
flask-login
email-validator
python-dotenv
flask_cors
numpy
pillow
mysqlclient
Flask-Markdown
psycopg2-binary
python-dotenv
flask_praetorian